# Issue 3774: Update the default max connection and Netty channel water marking for performance Improvement